openai_json/ml_processor.py
import logging
import joblib
from sklearn.ensemble import RandomForestClassifier  # Example model


class MachineLearningProcessor:
    """
    Uses a trained machine learning model to handle unmatched keys or structures.

    This class predicts schema-compliant transformations for unmatched keys
    and values using a pre-trained machine learning model.
    """

    def __init__(self, model_path: str = None):
        """
        Initializes the processor and optionally loads a pre-trained model.

        Args:
            model_path (str, optional): Path to the trained model file. If None, no model is loaded.
        """
        self.logger = logging.getLogger(__name__)
        self.model = None
        if model_path:
            self.load_model(model_path)

    def load_model(self, model_path: str):
        """
        Loads the machine learning model from a file.

        Args:
            model_path (str): Path to the trained model file.

        Raises:
            ValueError: If the model file cannot be loaded.
        """
        self.logger.info("Attempting to load model from path: %s", model_path)
        try:
            self.model = joblib.load(model_path)
            self.logger.info("Model loaded successfully from %s.", model_path)
        except Exception as e:
            self.logger.error("Failed to load model from %s: %s", model_path, e)
            raise ValueError(f"Failed to load model from {model_path}: {e}")

    def predict_transformations(self, unmatched_data: dict) -> dict:
        """
        Predicts schema-compliant transformations for unmatched keys and values.

        If no model is loaded, the unmatched data is returned unchanged.

        Args:
            unmatched_data (dict): A dictionary of keys and values that do not match the schema.

        Returns:
            dict: A dictionary of transformed keys and values compliant with the schema.
        """
        if not self.model:
            self.logger.warning("No model loaded. Exiting Maching Learning Processor.")
            return []

        self.logger.debug(
            "Predicting transformations for unmatched data: %s", unmatched_data
        )
        transformed_data = []

        # TODO Verify this implementation once the model is done.
        for key, value in unmatched_data.items():
            try:
                feature_vector = self._prepare_features(key, value)
                self.logger.debug(
                    "Generated feature vector for key '%s': %s", key, feature_vector
                )

                transformed_value = self.model.predict([feature_vector])[0]
                transformed_data[key] = transformed_value

                self.logger.debug(
                    "Predicted transformation for key '%s': original value '%s', transformed value '%s'.",
                    key,
                    value,
                    transformed_value,
                )
            except Exception as e:
                transformed_data[key] = f"Error processing {key}: {e}"
                self.logger.error(
                    "Error predicting transformation for key '%s' with value '%s': %s",
                    key,
                    value,
                    e,
                )

        self.logger.info("Transformation predictions completed: %s", transformed_data)
        return transformed_data

    def _prepare_features(self, key: str, value) -> list:
        """
        Converts a key-value pair into a feature vector suitable for the model.

        This method is used to extract features from unmatched data for prediction.
        The feature extraction logic can be customized based on the use case.

        Args:
            key (str): The unmatched key.
            value: The value associated with the unmatched key.

        Returns:
            list: A feature vector representing the key and value.
        """
        self.logger.debug("Preparing features for key '%s' and value '%s'.", key, value)
        # Example feature extraction (customize this as needed)
        feature_vector = [
            len(key),  # Length of the key
            len(str(value)),  # Length of the string representation of the value
            isinstance(value, (int, float)),  # Boolean: is the value numeric?
            isinstance(value, str),  # Boolean: is the value a string?
        ]
        self.logger.debug("Feature vector for key '%s': %s", key, feature_vector)
        return feature_vector
