tests/unit/test_ml_processor.py
from openai_json.ml_processor import MachineLearningProcessor
from sklearn.ensemble import RandomForestClassifier
import joblib
import pytest
from unittest.mock import patch, MagicMock


@pytest.fixture
def mock_model():
    mock = MagicMock()
    mock.predict.return_value = ["transformed_value"]
    return mock


def test_ml_processor_load_model(tmp_path):
    # Create a path where the model would be saved
    model_path = tmp_path / "model.pkl"

    # Assert that the model file does not exist (since it is not created yet)
    assert not model_path.exists()

    # Uncomment the following lines once a model is ready to test
    # joblib.dump(RandomForestClassifier(), model_path)
    # processor = MachineLearningProcessor()
    # processor.load_model(str(model_path))
    # assert processor.model is not None


@pytest.mark.skip(reason="ML functionality not yet implemented")
def test_ml_processor_predict(mock_model):
    processor = MachineLearningProcessor()
    processor.model = mock_model

    unmatched_data = {"extra_key": "extra_value"}
    transformed_data = processor.predict_transformations(unmatched_data)

    assert transformed_data == {"extra_key": "transformed_value"}


@pytest.mark.skip(reason="ML functionality not yet implemented")
@patch("openai_json.ml_processor.MachineLearningProcessor.load_model")
def test_mocked_load_model(mock_load_model):
    processor = MachineLearningProcessor()
    processor.load_model("mock-path")  # This call is mocked
    mock_load_model.assert_called_once_with("mock-path")
